+++ app/views/impact_cards/new.html.erb
@@ -171,21 +171,7 @@
             <h3 class="pl-10 text-sm font-semibold text-zinc-950 dark:text-white text-right">Subsystem Tags</h3>
             <p class="mt-1 text-sm/6 text-gray-600 text-right">Identify relevant subsystem tags for this initiative.</p>
           </div>
-          <div class="grid max-w-2xl grid-cols-1 gap-x-6 gap-y-8 sm:grid-cols-6 md:col-span-2" id="initiatives-subsystem-tags" data-controller="initiatives-subsystem-tags">
-            <%= render 'subsystem_tag_fields' %>
-            <div class="sm:col-span-6" data-initiatives-subsystem-tags-target="initiativesSubsystemTags">
-              <%= initiatives_form.fields_for :initiatives_subsystem_tags do |initiatives_subsystem_tags_form| %>
-                <div class="flex items-center space-x-4 mt-3">
-                  <%= initiatives_subsystem_tags_form.collection_select :subsystem_tag_id, policy_scope(SubsystemTag), :id, :name, { include_blank: 'Select a subsystem tag' } %>
-                  <%= link_to 'Remove', '#', class: 'text-red-500', data: { action: 'initiatives-subsystem-tags#removeTag' } %>
-                </div>
-              <% end %>
-            </div>
-
-            <div class="sm:col-span-6 mt-2">
-              <%= link_to 'Add Subsystem Tag', '#', id: 'add-subsystem-tag', class: 'inline-flex items-center px-4 py-2 border border-transparent text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500', data: { action: 'initiatives-subsystem-tags#addTag' } %>
-            </div>
-          </div>
+          <%= render 'initiatives_subsystem_tags/controls', form: initiatives_form %>
         </div>
       <% end %>
     <% end %>
