+++ app/views/initiatives/_form.html.erb
@@ -43,23 +43,7 @@
           <h3 class="pl-10 text-sm font-semibold text-zinc-950 dark:text-white text-right">Subsystem Tags</h3>
           <p class="mt-1 text-sm/6 text-gray-600 dark:text-gray-400 text-right">Identify relevant subsystem tags for this initiative.</p>
         </div>
-        <div class="grid max-w-2xl grid-cols-1 gap-x-6 gap-y-8 sm:grid-cols-6 md:col-span-2" id="initiatives-subsystem-tags" data-controller="initiatives-subsystem-tags">
-          <%= render 'subsystem_tag_fields' %>
-          <div class="sm:col-span-6" data-initiatives-subsystem-tags-target="initiativesSubsystemTags">
-            <%= turbo_frame_tag :subsystem_tags do %>
-              <%= form.fields_for :initiatives_subsystem_tags do |initiatives_subsystem_tags_form| %>
-                <div class="flex items-center space-x-4 mt-3 subsystem-tag">
-                  <%= initiatives_subsystem_tags_form.collection_select :subsystem_tag_id, policy_scope(SubsystemTag), :id, :name, { include_blank: 'Select a subsystem tag' } %>
-                  <%= link_to 'Remove', '#', class: 'text-red-500', data: { action: 'initiatives-subsystem-tags#removeTag' } %>
-                </div>
-              <% end %>
-            <% end %>
-          </div>
-
-          <div class="sm:col-span-6 mt-2">
-            <%= render 'initiatives/subsystem_tags/index', initiative: %>
-          </div>
-        </div>
+        <%= render 'initiatives_subsystem_tags/controls', form: %>
       </div>
 
       <div class="grid grid-cols-1 gap-x-8 gap-y-10 border-b border-gray-900/10 dark:border-white/10 pb-12 md:grid-cols-3">
