--- app/views/initiatives/_subsystem_tag_fields.html.erb
@@ -1,12 +1,0 @@
-<template id="subsystem_tag_template">
-  <div class="flex items-center space-x-4 mt-2">
-    <%=
-      custom_select_tag(
-        'initiative[initiatives_subsystem_tags_attributes][__INDEX__][subsystem_tag_id]',
-        options_from_collection_for_select(policy_scope(SubsystemTag), :id, :name),
-        include_blank: 'Select a subsystem tag'
-      )
-    %>
-    <%= link_to 'Remove', '#', class: 'text-red-500', data: { action: 'initiatives-subsystem-tags#removeTag' } %>
-  </div>
-</template>
