--- app/views/initiatives/subsystem_tags/new.html.erb
@@ -1,3 +1,0 @@
-<%= turbo_frame_tag :new_subsystem_tag do %>
-  <%= render 'form', label: @subsystem_tag%>
-<% end %>
