+++ app/views/initiatives_subsystem_tags/_controls.html.erb
@@ -1,0 +1,25 @@
+<div class="grid max-w-2xl grid-cols-1 gap-x-6 gap-y-8 sm:grid-cols-6 md:col-span-2" id="initiatives-subsystem-tags" data-controller="initiatives-subsystem-tags">
+  <%= render 'initiatives_subsystem_tags/subsystem_tag_fields' %>
+  <div class="sm:col-span-6" data-initiatives-subsystem-tags-target="initiativesSubsystemTags">
+  <%= turbo_frame_tag :subsystem_tags do %>
+      <%= form.fields_for :initiatives_subsystem_tags do |initiatives_subsystem_tags_form| %>
+      <div class="flex items-center space-x-4 mt-3 subsystem-tag">
+          <%= 
+            initiatives_subsystem_tags_form.collection_select(
+              :subsystem_tag_id, 
+              policy_scope(SubsystemTag), 
+              :id, 
+              :name, 
+              { include_blank: 'Select a subsystem tag' }
+            ) 
+          %>
+          <%= link_to 'Remove', '#', class: 'text-red-500', data: { action: 'initiatives-subsystem-tags#removeTag' } %>
+      </div>
+      <% end %>
+  <% end %>
+  </div>
+
+  <div class="sm:col-span-6 mt-2">
+  <%= render 'initiatives_subsystem_tags/index' %>
+  </div>
+</div>
