+++ app/views/initiatives_subsystem_tags/_form.html.erb
@@ -1,5 +1,5 @@
 <div data-controller="labels">
-  <%= form_with(model: [@initiative, @subsystem_tag], html: { class: 'form-horizontal record-form border-b border-zinc-950/10 dark:border-white/10 pb-5' }) do |f| %>
+  <%= form_with(model: @subsystem_tag, url: initiatives_subsystem_tags_path, html: { class: 'form-horizontal record-form border-b border-zinc-950/10 dark:border-white/10 pb-5' }) do |f| %>
     <div class="w-full mb-4">
         <div class="flex items-center justify-between py-4">
           <div class="flex items-center gap-6 w-1/3">
@@ -45,7 +45,7 @@
         <%= 
           link_to(
             "Cancel", 
-            initiative_subsystem_tags_path(initiative_id: params[:initiative_id]), 
+            initiatives_subsystem_tags_path, 
             class: LinksHelper::SECONDARY_CLASS,
             data: { turbo_stream: true } 
           )
