+++ app/views/initiatives_subsystem_tags/create.turbo_stream.erb
@@ -1,4 +1,4 @@
-<turbo-stream action="update" target="new_subsystem_tag">
+<turbo-stream action="update" target="new_initiatives_subsystem_tag">
   <template>
     <%= 
       link_to(
@@ -13,9 +13,9 @@
     <%= 
       link_to(
         'Create Subsystem Tag', 
-        new_initiative_subsystem_tag_path(initiative_id: params[:id] || params[:initiative_id]), 
+        new_initiatives_subsystem_tag_path, 
         class: 'ml-2 inline-flex items-center px-4 py-2 border border-transparent text-sm font-medium rounded-md text-white bg-emerald-600 hover:bg-emerald-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-emerald-500', 
-        data: { turbo_frame: 'new_subsystem_tag' }
+        data: { turbo_frame: 'new_initiatives_subsystem_tag' }
       )
     %>
   </template>
@@ -23,7 +23,7 @@
 
 <turbo-stream action="append" target="subsystem_tags">
   <template>
-    <div class="flex items-center space-x-4 mt-2">
+    <div class="flex items-center space-x-4 mt-2 subsystem-tag">
       <%# TODO: Replace this index 'hack' with an index value in sequence with tags on the page %>
       <% index = @subsystem_tag.id + rand(100000) %>
       <%=
