+++ app/views/initiatives_subsystem_tags/new.html.erb
@@ -1,0 +1,3 @@
+<%= turbo_frame_tag :new_initiatives_subsystem_tag do %>
+  <%= render 'form', label: @subsystem_tag%>
+<% end %>
